{"schema":"decodex/commit/1","summary":"Split service acceptance test modules","authority":"manual"}
